+++ plot_antenna/config.py
@@ -3,7 +3,7 @@ Configuration file for the RFlect application.
 
 This module contains constants and configuration values used throughout the application.
 """
-
+interpolate_3d_plots = True  # Default value, can be set to False to disable interpolation
 # 3D Plotting Interpolation Resolution for viewing plots (lower for better performance)
 PHI_RESOLUTION = 120 #default 360 = 1deg spacing
 THETA_RESOLUTION = 60 #default 180 - 1deg spacing
@@ -28,3 +28,4 @@ ACCENT_GREEN_COLOR = "#4CAF50"
 # Fonts
 HEADER_FONT = ("Arial", 14, "bold")
 LABEL_FONT = ("Arial", 12)
+